# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ADYPU Grounded Answer Engine API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Admin login
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    username: zod.string(),
    role: zod.string(),
  }),
});

/**
 * @summary Ask a question
 */
export const chatBodyQuestionMax = 2000;

export const ChatBody = zod.object({
  question: zod.string().min(1).max(chatBodyQuestionMax),
  sessionId: zod.string().optional(),
});

export const ChatResponse = zod.object({
  answer: zod.string(),
  intent: zod.string(),
  entities: zod.array(
    zod.object({
      type: zod.string(),
      value: zod.string(),
    }),
  ),
  citations: zod.array(
    zod.object({
      chunkId: zod.number(),
      documentTitle: zod.string(),
      sourceUrl: zod.string(),
      excerpt: zod.string(),
      relevanceScore: zod.number(),
    }),
  ),
  confidence: zod.object({
    score: zod.number(),
    label: zod.enum(["High", "Medium", "Low"]),
    breakdown: zod
      .object({
        retrieval: zod.number().optional(),
        rerank: zod.number().optional(),
        sourceAgreement: zod.number().optional(),
        freshness: zod.number().optional(),
        answerability: zod.number().optional(),
      })
      .optional(),
  }),
  queryLogId: zod.number().optional(),
  rewrittenQuery: zod.string().optional(),
});

/**
 * @summary Start a new crawl job
 */
export const startCrawlBodyMaxDepthDefault = 3;

export const StartCrawlBody = zod.object({
  sourceIds: zod.array(zod.number()).optional(),
  maxDepth: zod.number().default(startCrawlBodyMaxDepthDefault),
});

export const StartCrawlResponse = zod.object({
  jobId: zod.number(),
  status: zod.string(),
  message: zod.string().optional(),
});

/**
 * @summary Recrawl stale content
 */
export const RecrawlStaleResponse = zod.object({
  jobId: zod.number(),
  status: zod.string(),
  message: zod.string().optional(),
});

/**
 * @summary Get admin dashboard stats
 */
export const GetAdminStatsResponse = zod.object({
  totalDocuments: zod.number(),
  totalChunks: zod.number(),
  failedPages: zod.number(),
  totalSources: zod.number(),
  activeJob: zod
    .object({
      id: zod.number(),
      status: zod.string(),
      pagesFound: zod.number(),
      pagesProcessed: zod.number(),
      pagesFailed: zod.number(),
      startedAt: zod.string().optional(),
      completedAt: zod.string().optional(),
    })
    .optional(),
  recentJobs: zod.array(
    zod.object({
      id: zod.number(),
      status: zod.string(),
      pagesFound: zod.number(),
      pagesProcessed: zod.number(),
      pagesFailed: zod.number(),
      startedAt: zod.string().optional(),
      completedAt: zod.string().optional(),
    }),
  ),
  lowConfidenceCount: zod.number(),
});

/**
 * @summary Get indexed documents
 */
export const getAdminDocumentsQueryPageDefault = 1;
export const getAdminDocumentsQueryLimitDefault = 20;

export const GetAdminDocumentsQueryParams = zod.object({
  page: zod.coerce.number().default(getAdminDocumentsQueryPageDefault),
  limit: zod.coerce.number().default(getAdminDocumentsQueryLimitDefault),
  source: zod.coerce.string().optional(),
  status: zod.coerce.string().optional(),
});

export const GetAdminDocumentsResponse = zod.object({
  documents: zod.array(
    zod.object({
      id: zod.number(),
      title: zod.string(),
      sourceUrl: zod.string(),
      contentType: zod.string(),
      chunkCount: zod.number(),
      createdAt: zod.string().optional(),
      updatedAt: zod.string().optional(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
});

/**
 * @summary Get unanswered or low-confidence questions
 */
export const getLowConfidenceQuestionsQueryPageDefault = 1;
export const getLowConfidenceQuestionsQueryLimitDefault = 20;

export const GetLowConfidenceQuestionsQueryParams = zod.object({
  page: zod.coerce.number().default(getLowConfidenceQuestionsQueryPageDefault),
  limit: zod.coerce
    .number()
    .default(getLowConfidenceQuestionsQueryLimitDefault),
});

export const GetLowConfidenceQuestionsResponse = zod.object({
  questions: zod.array(
    zod.object({
      id: zod.number(),
      question: zod.string(),
      intent: zod.string(),
      confidenceScore: zod.number(),
      confidenceLabel: zod.string(),
      createdAt: zod.string(),
      hasAnswer: zod.boolean(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
});

/**
 * @summary Get allowed crawl sources
 */
export const GetAdminSourcesResponse = zod.object({
  sources: zod.array(
    zod.object({
      id: zod.number(),
      domain: zod.string(),
      urlPattern: zod.string(),
      isActive: zod.boolean(),
      createdAt: zod.string().optional(),
    }),
  ),
});

/**
 * @summary Add an allowed crawl source
 */
export const AddSourceBody = zod.object({
  domain: zod.string(),
  urlPattern: zod.string(),
  seedUrl: zod.string().optional(),
});

export const AddSourceResponse = zod.object({
  id: zod.number(),
  domain: zod.string(),
  urlPattern: zod.string(),
  isActive: zod.boolean(),
  createdAt: zod.string().optional(),
});

/**
 * @summary Remove an allowed crawl source
 */
export const RemoveSourceParams = zod.object({
  id: zod.coerce.number(),
});

export const RemoveSourceResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Submit feedback on an answer
 */
export const submitFeedbackBodyRatingMax = 5;

export const SubmitFeedbackBody = zod.object({
  queryLogId: zod.number(),
  rating: zod.number().min(1).max(submitFeedbackBodyRatingMax),
  comment: zod.string().optional(),
});

export const SubmitFeedbackResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});